requirements.txt
# Core dependencies
flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Production ready
requests==2.31.0

# Add your project dependencies below
# pandas==2.1.4
# numpy==1.26.2
